network:util: Fix warnings due invalid types
./src/lib/comms/sol-network-util.h: In function 'sol_bluetooth_is_addr_str':
/zephyr-sdk/sysroots/i586-poky-elf/usr/i586-poky-elf/include/ctype.h:57:54:
warning: array subscript has type 'char' [-Wchar-subscripts]
 #define __ctype_lookup(__c)
 ((__ctype_ptr__+sizeof(""[__c]))[(int)(__c)])

+++ src/lib/comms/sol-network-util.h
@@ -55,10 +55,10 @@ sol_bluetooth_is_addr_str(const char *str)
     const char *p = str;
 
     while (*str) {
-        if (!isxdigit(*str++))
+        if (!isxdigit((int)*str++))
             return false;
 
-        if (!isxdigit(*str++))
+        if (!isxdigit((int)*str++))
             return false;
 
         if (*str == '0')
